requirements.txt
BeautifulSoup==3.2.1
python-memcached==1.57
requests==2.8.1
six==1.10.0
wsgiref==0.1.2
